requirements.txt
Django==1.9.4
Fabric3==1.10.2.post3
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.23
cffi==1.6.0
cryptography==1.3.2
dj-database-url==0.4.1
django-celery==3.1.17
gunicorn==19.4.5
idna==2.1
kombu==3.0.35
paramiko==2.0.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pytz==2016.4
redis==2.10.5
requests==2.9.1
six==1.10.0